+++ nameko_sqlalchemy/transaction_retry.py
@@ -1,23 +1,24 @@
-
-import eventlet
 import functools
 import operator
 
 import wrapt
 from sqlalchemy import exc
+from nameko.utils.retry import retry
 
 
-def transaction_retry(wrapped=None, session=None, retries=1, backoff_factor=0):
+def transaction_retry(wrapped=None, session=None,
+                      max_attempts=1, delay=0, backoff=1, max_delay=None):
 
     if wrapped is None:
         return functools.partial(
-            transaction_retry, session=session,
-            retries=retries, backoff_factor=backoff_factor)
+            transaction_retry, session=session, max_attempts=max_attempts,
+            delay=delay, backoff=backoff, max_delay=max_delay)
 
     @wrapt.decorator
     def wrapper(wrapped, instance, args, kwargs):
 
-        @retry(retries, backoff_factor, exc.OperationalError)
+        @retry(for_exceptions=exc.OperationalError, max_attempts=max_attempts,
+               delay=delay, backoff=backoff, max_delay=max_delay)
         def run_or_rollback():
             try:
                 return wrapped(*args, **kwargs)
@@ -32,30 +33,3 @@ def transaction_retry(wrapped=None, session=None, retries=1, backoff_factor=0):
         return run_or_rollback()
 
     return wrapper(wrapped)  # pylint: disable=E1120
-
-
-def retry(retries, backoff_factor, exceptions):
-
-    retries = max(retries, 1)
-    backoff_factor = max(backoff_factor, 0)
-
-    @wrapt.decorator
-    def wrapper(wrapped, instance, args, kwargs):
-        errors = 0
-        while True:
-            try:
-                return wrapped(*args, **kwargs)
-            except exceptions:
-                errors += 1
-
-                if errors > retries:
-                    raise
-
-                if errors >= 2:
-                    backoff_value = backoff_factor * (2 ** (errors - 2))
-                else:
-                    backoff_value = 0
-
-                eventlet.sleep(backoff_value)
-
-    return wrapper